Move new request handling logic into tracessh client.
